fix try for progress bossbar and reduce spawn time (15 min -> 14 min)

+++ src/main/java/fr/openmc/core/features/events/contents/dailyevents/DailyEventsManager.java
@@ -242,7 +242,7 @@ public class DailyEventsManager extends Feature implements LoadAfterItemsAdder, 
      * @param event l'event
      * @return le temps restant en secondes
      */
-    public static double getRemainingTime(DailyEvent event) {
+    public static float getRemainingTime(DailyEvent event) {
         if (!isActiveDailyEvent()) return -1;
         if (getActiveDailyEvent() != event) return -1;
